backend/requirements.txt
fastapi
uvicorn
Pillow
ImageHash
supabase
python-multipart
